+++ src/complete_gen.rs
@@ -403,7 +403,10 @@ impl State {
             7 => render_zsh(&items, &shell, full_lit),
             8 => render_bash(&items, &shell, full_lit),
             9 => render_fish(&items, &shell, full_lit, self.path[0].as_str()),
-            unk => panic!("Unsupported output revision {}, you need to genenerate your shell completion files for the app", unk)
+            unk => {
+                eprintln!("Unsupported output revision {}, you need to genenerate your shell completion files for the app", unk);
+                std::process::exit(1);
+            }
         }.unwrap())
     }
 }