+++ Changelog.md
@@ -6,6 +6,7 @@
 - better error messages for unexpected values that prevent positional parses
 - bugfix in completions generator for bash
   thanks @akinomyoga
+- `choice` combinator to efficiently construct alternative parsers at runtime
 
 ## bpaf [0.9.5], bpaf_derive [0.5.5] - 2023-08-24
 - fancier squashing: parse `-abfoo` as `-a -b=foo` if b is a short argument