+++ Changelog.md
@@ -1,9 +1,12 @@
 # Change Log
 
-## bpaf [0.9.6], bpaf_derive [0.5.6] - Unreleased
+## bpaf [0.9.6], bpaf_derive [0.5.6] - 2023-10-30
 - make sure env-only arguments and flags are working
 - support raw identifiers in derive macro (#282)
 - better error messages for unexpected values that prevent positional parses
+- bugfix in completions generator for bash
+  thanks @akinomyoga
+- `choice` combinator to efficiently construct alternative parsers at runtime
 
 ## bpaf [0.9.5], bpaf_derive [0.5.5] - 2023-08-24
 - fancier squashing: parse `-abfoo` as `-a -b=foo` if b is a short argument