+++ Changelog.md
@@ -1,6 +1,25 @@
 # Change Log
 
 
+## bpaf [0.10.1] - Unreleased
+- all the parsers are moved under SimpleParser
+- `literal` parser now takes an extra parameter to be closer to `req_flag`
+- removed standalone `command`
+
+### Migration guide 0.9.x -> 0.10.0
+- `literal` parser now takes an extra parameter it will produce to be closer to `req_flag`
+  ```diff
+  -let turbo = literal("+turbo").map(|_| true).fallback(false);
+  +let turbo = literal("+turbo", true).fallback(false);
+  ```
+  To replicate old behavior exactly you can pass `()` as a second parameter
+
+- standalone `command` function is gone, you should replace it with a member on OptionParser:
+  ```diff
+  -let a = command("name", subparser);
+  +let a = subparser.command("name");
+  ```
+
 ## bpaf - unreleased
 - fix formatting in ambiguity error message
 - drop batteries::toggle_flag